+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 *[CalVer, YY.month.patch](https://calver.org/)*
 
 ## Future
+- Added TRIOXXX check: Async functions must have at least one checkpoint on every code path, unless an exception is raised
 - Add TRIO103: `except BaseException` or `except trio.Cancelled` with a code path that doesn't re-raise
 - Add TRIO104: "Cancelled and BaseException must be re-raised" if user tries to return or raise a different exception.